+++ iOSAudio/iOSAudio/AudioRecorderManager.swift
@@ -45,8 +45,8 @@ final class AudioRecorderManager: NSObject {
     private(set) var status: EngineStatus = .notInitialized
     private(set) var streamingInProgress: Bool = false
 
-    init(sampleRate: Int = 16000, numberOfChannels: UInt32 = 1, commonFormat: AVAudioCommonFormat = .pcmFormatInt16) {
-        self.sampleRate = Double(sampleRate)
+    init(sampleRate: Double = 16000, numberOfChannels: UInt32 = 1, commonFormat: AVAudioCommonFormat = .pcmFormatInt16) {
+        self.sampleRate = sampleRate
         self.numberOfChannels = numberOfChannels
         self.commonFormat = commonFormat
     }
